change strlen return type

+++ str_utils.c
@@ -5,7 +5,7 @@
  * @str: string
  * Return: length of string
  */
-int _strlen(const char *str)
+size_t _strlen(const char *str)
 {
 	int i = 0;
